Add osm plot unittests
