+++ src/wktplot/plots/osm.py
@@ -2,7 +2,7 @@ from bokeh import plotting as plt
 from bokeh.tile_providers import Vendors, get_provider
 from typing import Any, Dict
 from wktplot.plots.standard import WKTPlot
-from wktplot.maps import OpenStreetMap
+from wktplot.maps.osm import OpenStreetMap
 
 
 class OpenStreetMapsPlot(WKTPlot):
@@ -10,6 +10,10 @@ class OpenStreetMapsPlot(WKTPlot):
     """
 
     mapper = OpenStreetMap
+    default_figure_style_kwargs: Dict[str, str] = {
+        "x_axis_type": "mercator",
+        "y_axis_type": "mercator",
+    }
 
     @classmethod
     def _create_figure(cls, title: str, **style_kwargs: Dict[str, Any]) -> plt.Figure:
@@ -18,12 +22,11 @@ class OpenStreetMapsPlot(WKTPlot):
         # - https://docs.bokeh.org/en/latest/docs/user_guide/geo.html#tile-provider-maps
         default_kwargs: Dict[str, Any] = {
             **style_kwargs,
-            "x_axis_type": "mercator",
-            "y_axis_type": "mercator",
+            **cls.default_figure_style_kwargs,
         }
 
         tile_provider = get_provider(Vendors.OSM)
-        fig = super().create_figure(title=title, **default_kwargs)
+        fig = super()._create_figure(title=title, **default_kwargs)
         fig.add_tile(tile_provider)
 
         return fig
