+++ tests/unit/plots/common.py
@@ -1,0 +1,3 @@
+PLOT_TITLE = "Michael Plot ABC123"
+PLOT_FILE = "michael_plot_abc123.html"
+STYLE_KWARGS = {"color": "MidnightBlue", "line_width": 3.0}
